build: remove pdoc from poetry dependencies
pdoc is only used to build the documentation website and/or wiki pages. However, a different version of pdoc is
used for each (pdoc vs pdoc3). Because both of these processes take place via GitHub Actions, the correct pdoc
version can be installed in the Action itself
starting # with '#' will be ignored, and an empty message aborts the commit.